Clarify that minimum viewport scale may change
The minimum viewport scale must ensure that the viewport has a nonzero
size, and this may in theory require an adjustment if the session's
baseLayer is changed during a session to use a very small
framebufferScaleFactor.

Change the wording to clarify that it's not necessarily constant for an
entire session.